New Plugin Processing Logic

+++ server/plugins/index.js
@@ -15,7 +15,38 @@ const fs = require('fs');
 const latestVersion = require('latest-version');
 const npm = require('npm-programmatic');
 const sequelize = require('sequelize');
+const _eval = require('eval');
+const toolsLib = require('hollaex-tools-lib');
 const lodash = require('lodash');
+const expressValidator = require('express-validator');
+const multer = require('multer');
+const moment = require('moment');
+const mathjs = require('mathjs');
+const bluebird = require('bluebird');
+const umzug = require('umzug');
+const rp = require('request-promise');
+const uuid = require('uuid/v4');
+const jwt = require('jsonwebtoken');
+const momentTz = require('moment-timezone');
+const json2csv = require('json2csv');
+const flat = require('flat');
+const ws = require('ws');
+const cron = require('node-cron');
+const randomString = require('random-string');
+const bcryptjs = require('bcryptjs');
+const expectCt = require('expect-ct');
+const validator = require('validator');
+const otp = require('otp');
+const geoipLite = require('geoip-lite');
+const nodemailer = require('nodemailer');
+const wsHeartbeatServer = require('ws-heartbeat/server');
+const wsHeartbeatClient = require('ws-heartbeat/client');
+const winston = require('winston');
+const elasticApmNode = require('elastic-apm-node');
+const winstonElasticsearchApm = require('winston-elasticsearch-apm');
+const tripleBeam = require('triple-beam');
+const uglifyEs = require('uglify-es');
+const bodyParser = require('body-parser');
 const pluginProcess = path.join(__dirname, "./plugin-process.js");
 const { Worker } = require('worker_threads')
 const { createProxyMiddleware } = require('http-proxy-middleware');
@@ -84,8 +115,16 @@ const stopPlugin = async (plugin) => {
 			`killing plugin ${plugin.name}`
 		);
 
-		activePlugins[plugin.name].process.terminate();
-		delete activePlugins[plugin.name];
+		if (activePlugins[plugin.name]) {
+			activePlugins[plugin.name].process.terminate();
+			activePlugins[plugin.name].active = false;
+		} else {
+			activePlugins[plugin.name] =
+			{
+				name: plugin.name,
+				active: false
+			}
+		}
 
 	} catch (err) {
 		loggerPlugin.error(
@@ -127,7 +166,8 @@ const startPlugin = async (plugin) => {
 		activePlugins[plugin.name] = {
 			process: childProcess,
 			port: pluginData.PORT,
-			name: plugin.name
+			name: plugin.name,
+			active: true
 		};
 
 
@@ -162,13 +202,12 @@ checkStatus()
 		app.use(logEntryRequest);
 		app.use(domainMiddleware);
 		helmetMiddleware(app);
-
-
 		const defaultURL = 'http://localhost:10012';
+
 		const customRouter = function (req) {
 			if (req.path.length > 1 && req.path.includes('/plugins/')) {
 				for (let plugin of Object.values(activePlugins)) {
-					if (req.path.includes(plugin.name)) {
+					if (req.path.includes(plugin.name) && plugin.active) {
 						return `http://localhost:${plugin.port}`;
 					}
 				}
@@ -182,8 +221,25 @@ checkStatus()
 			router: customRouter,
 			changeOrigin: true
 		};
+		const serviceProviderProxy = createProxyMiddleware(options);
 
-		app.use('/plugins', routes, createProxyMiddleware(options));
+		app.use((req, res, next) => {
+			if (req.path.length > 1 && req.path.includes('/plugins/')) {
+				for (let plugin of Object.values(activePlugins)) {
+					if (req.path.includes(plugin.name)) {
+						if (plugin.active) {
+							return serviceProviderProxy.call(serviceProviderProxy, req, res, next);
+						} else {
+							return res.status(404).send();
+						}
+					}
+				}
+			}
+			next();
+		});
+
+
+		app.use('/plugins', routes);
 
 		const plugins = await Plugin.findAll({
 			where: {
@@ -196,7 +252,93 @@ checkStatus()
 		});
 
 		for (const plugin of plugins) {
-			startPlugin(plugin);
+			try {
+				loggerPlugin.verbose(
+					'plugins/index/initialization',
+					`starting plugin ${plugin.name}`
+				);
+
+				const context = {
+					configValues: {
+						publicMeta: plugin.public_meta,
+						meta: plugin.meta
+					},
+					pluginLibraries: {
+						app,
+						loggerPlugin,
+						toolsLib
+					},
+					app,
+					toolsLib,
+					lodash,
+					expressValidator,
+					loggerPlugin,
+					multer,
+					moment,
+					mathjs,
+					bluebird,
+					umzug,
+					rp,
+					sequelize,
+					uuid,
+					jwt,
+					momentTz,
+					json2csv,
+					flat,
+					ws,
+					cron,
+					randomString,
+					bcryptjs,
+					expectCt,
+					validator,
+					uglifyEs,
+					otp,
+					latestVersion,
+					geoipLite,
+					nodemailer,
+					wsHeartbeatServer,
+					wsHeartbeatClient,
+					cors,
+					winston,
+					elasticApmNode,
+					winstonElasticsearchApm,
+					tripleBeam,
+					bodyParser,
+					morgan,
+					meta: plugin.meta,
+					publicMeta: plugin.public_meta,
+					installedLibraries: {}
+				};
+
+				if (plugin.prescript && lodash.isArray(plugin.prescript.install) && !lodash.isEmpty(plugin.prescript.install)) {
+					loggerPlugin.verbose(
+						'plugins/index/initialization',
+						`Installing packages for plugin ${plugin.name}`
+					);
+
+					for (const library of plugin.prescript.install) {
+						context.installedLibraries[library] = await installLibrary(library);
+					}
+
+					loggerPlugin.verbose(
+						'plugins/index/initialization',
+						`Plugin ${plugin.name} packages installed`
+					);
+				}
+
+				_eval(plugin.script, plugin.name, context, true);
+
+				loggerPlugin.verbose(
+					'plugins/index/initialization',
+					`Plugin ${plugin.name} running`
+				);
+			} catch (err) {
+				loggerPlugin.error(
+					'plugins/index/initialization',
+					`error while starting plugin ${plugin.name}`,
+					err.message
+				);
+			}
 		}
 
 		loggerPlugin.info(
